Foundation: use NS(U)Integer where appropriate; use OBJC_MALLOC() or at least add a comment; add objc_check_malloc() every now and then to be able to debug heap corruption; other small fixes

+++ Foundation/Sources/NSAutoreleasePool.h
@@ -33,7 +33,7 @@
 	struct autorelease_array_list *_released;
 	struct autorelease_array_list *_released_head;
 
-	unsigned _released_count;	// number of objects autoreleased in this pool
+	NSUInteger _released_count;	// number of objects autoreleased in this pool
 }
 
 + (void) addObject:(id) anObject;
